# backend/app/app/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: GoWind Content Hub App API
    description: GoWind Content Hub App API
    contact:
        name: tx7do
        url: https://github.com/tx7do/go-wind-ledger
        email: yanglinbo@gmail.com
    license:
        name: MIT License
        url: https://github.com/tx7do/go-wind-ledger/blob/master/LICENSE
    version: "1.0"
paths:
...
    /app/v1/flow-files:
        get:
            tags:
                - FlowFileService
            operationId: FlowFileService_List
            parameters:
                - name: flowId
                  in: query
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.ListFlowFileResponse'
    /app/v1/flow-files-upload:
        post:
            tags:
                - FlowFileService
            operationId: FlowFileService_UploadFile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ledger.service.v1.UploadFlowFileRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.FlowFile'
    /app/v1/flow-files-view:
        get:
            tags:
                - FlowFileService
            operationId: FlowFileService_ViewFile
            parameters:
                - name: id
                  in: query
                  schema:
                    type: integer
                    format: uint32
                - name: createTime
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ledger.service.v1.ViewFlowFileResponse'
    /app/v1/flow-files/{id}:
        delete:
            tags:
                - FlowFileService
            operationId: FlowFileService_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: uint32
            responses:
                "200":
                    description: OK
                    content: {}
...
components:
    schemas:
...
        ledger.service.v1.FlowFile:
            type: object
            properties:
                id:
                    type: integer
                    format: uint32
                flowId:
                    type: integer
                    description: 关联的流水ID
                    format: uint32
                creatorId:
                    type: integer
                    format: uint32
                createTime:
                    type: string
                contentType:
                    type: string
                size:
                    type: string
                originalName:
                    type: string
            description: 流水附件
...
        ledger.service.v1.ListFlowFileResponse:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/ledger.service.v1.FlowFile'
            description: 回应 - 附件列表
...
        ledger.service.v1.UploadFlowFileRequest:
            type: object
            properties:
                flowId:
                    type: integer
                    format: uint32
                fileName:
                    type: string
                contentType:
                    type: string
                size:
                    type: string
                data:
                    type: string
                    format: bytes
            description: 请求 - 上传附件
        ledger.service.v1.ViewFlowFileResponse:
            type: object
            properties:
                contentType:
                    type: string
                originalName:
                    type: string
                data:
                    type: string
                    format: bytes
            description: 回应 - 查看附件（返回文件内容）
...
    securitySchemes:
        OAuth2PasswordBearer:
            type: oauth2
            flows:
                password:
                    tokenUrl: /admin/v1/login
                    refreshUrl: /admin/v1/refresh-token
                    scopes: {}
security:
    - OAuth2PasswordBearer: []
...